+++ server/app/boto3_handlers/common_funcs.py
