+++ server/app/boto3_handlers/kloud_resource_client.py
@@ -1,0 +1,6 @@
+import boto3
+
+
+class KloudResourceClient:
+    def __init__(self, session_instance: boto3.Session):
+        self._session = session_instance
